RavenDB-10648 Javascript Map-Reduce ---!!!WIP!!!---
