+++ src/Raven.Server/Documents/Indexes/Persistence/Lucene/Documents/IPropertyAccessor.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Raven.Server.Documents.Indexes.Persistence.Lucene.Documents
+{
+    public interface IPropertyAccessor
+    {
+        IEnumerable<(string Key, object Value, bool IsGroupByField)> GetPropertiesInOrder(object target);
+
+        object GetValue(string name, object target);
+    }
+}
