+++ test/FastTests/Client/Indexing/JavaScriptIndexTests.cs
@@ -148,7 +148,7 @@ namespace FastTests.Client.Indexing
 
             }
         }
-        [Fact(Skip = "The code for this is not yet ready")]
+        [Fact]
         public void CanUseJavaScriptMapReduceIndex()
         {
             using (var store = GetDocumentStore())
@@ -167,6 +167,7 @@ namespace FastTests.Client.Indexing
                         IsAvailable = true
                     });
                     session.SaveChanges();
+                    WaitForUserToContinueTheTest(store);
                     WaitForIndexing(store);
                     session.Query<User>("UsersAndProductsByNameAndCount").OfType<ReduceResults>().Single(x => x.Name == "Brendan Eich");
                 }
@@ -323,7 +324,7 @@ map('Users', function (u){
                             @"map('Users', function (u){ return { Name: u.Name, Count: 1};})",
                             @"map('Products', function (p){ return { Name: p.Name, Count: 1};})"
                         },
-                    Reduce = @"groupBy(x => x.Name).aggregate((key,values) => {return {Name: key,Count: values.reduce((total, val) => val.Count + total,0)};})",
+                    Reduce = @"groupBy(function(x) { return x.Name;}).aggregate((key,values) => {return {Name: key,Count: values.reduce((total, val) => val.Count + total,0)};})",
                     Type = IndexType.JavaScriptMapReduce,
                     LockMode = IndexLockMode.Unlock,
                     Priority = IndexPriority.Normal,
